feat: ocr 텍스트 값과 opencv 색깔 블럭 매칭 코드 구현

+++ server/Whatta/src/main/java/whatta/Whatta/ocr/payload/dto/BoundingPoly.java
@@ -10,7 +10,7 @@ public record BoundingPoly(
 ) {
     @Builder
     public record Vertex(
-            double x,
-            double y
+            int x,
+            int y
     ) {}
 }
